// src/pages/book/Interface.ts
export interface ICreateOrderRequestType {
  /**
   * 联系人信息
   */
  contactInfo?:
    {
        /**
         * 全名
         */
        contactName: string | null;
        /**
         * 邮箱
         */
        email: string | null;
        /**
         * 电话地区编号
         */
        phoneArea: string | null;
        /**
         * 联系电话
         */
        contactTel: string | null;
        /**
         * 手机号码
         */
        mobilePhone: string | null;
    };
  /**
   * 乘机人信息
   */
  flightPassengerList:
    [
        {
          /**
           * 乘机人ID
           */
          passengerId?: string | null;
          /**
           * 乘机人名
           */
          givenName: string | null;
          /**
           * 乘机人姓
           */
          surName: string | null;
          /**
           * 生日
           */
          birthDay: string | null;
          /**
           * 证件号
           */
          cardNo?: string | null;
          /**
           * 乘客类型 ADT/CHD/INF
           */
          travelerType: string | null;
          /**
           * 证件类型ID
           */
          cardType?: number | null;
          /**
           * 性别
           */
          gender: string | null;
          /**
           * 国家编号
           */
          nationality: string | null;
          /**
           * 证件有效期
           */
          passportLimit?: string | null;
        }
      ]
    | [];
  shoppingId: string | null;
  mktportal: string | null;
  remark: string | null;
  locale: string | null;
  channel: string;
  revenueId: string | null;
  profitId: string | null;
  currencyRate: string | null;
}